Make PDMats generic so that it works with other floating point types (in principle, any type supported by Base.LinAlg.Cholesky, which, dependend on the Julia version is Float64, Float32 and BigFloat)
Some problems with downstream sparse libraries mean that there are several workarounds
for PDSparseMat (documented in the code).

PDSparseMat with Float32 will probably only work with Julia 0.4.2 or above
because of julia issue #14076 (limitations in CHOLMOD library) for which
the commit of #14108 provides a workaround. A backport for that issue has
been submitted for Julia 0.4, according to the notes.

The tests for this commit are extended to cover both Float64 and Float32 cases.

Support for Float16 and BigFloat is partial (i.e., works for full matrices
from Julia 0.4, but doesn't work for sparse matrices, dependend on whether
the Cholesky factorization supports the type)

Changed Real constants to AbstractFloat (to be consistent with other
functions)

Made some small changes to the tests so that Float32 is not tested for
Julia 0.3 (it can't work because of limitations of cholesky functions).

Package should still work for 0.3 for Float 64, so changing requirements
back to let Travis run the tests. For Float32, it can't work.

Changed calls of T() to convert(T,...) to pass tests for Float64 in Julia
0.3.12. Even for Float32, all tests can pass except PDSparseMat (because
cholesky functions do not support Float32)

Fixed type parameters for all concrete types such that the external constructors are all
non-generic (this could have given problems in packages using PDMats such
as Distributions).

I added tests for all external constructors as some were missing.

Simplified the type parameters of PDMat and PDSparseMat such that the
Cholesky type is not part of the type definition (given that the type
parameters of Cholesky equal the two remaining type parameters S,T

Made constructor arguments consistent across the four different types of
PDMat.

Updated documentation to support new functionality

Additional edits to documentation page.

Minor edits to documentation